requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
openai>=1.6.1
python-dotenv==1.0.0
fake-useragent==1.4.0
flask==3.0.0
flask-cors==4.0.0
lxml==4.9.3
urllib3==2.0.7

